+++ src/vault/internal/folder/response.go
@@ -5,8 +5,8 @@ import (
 )
 
 type Response struct {
-	ID        uint      `json:"id"`
-	UserID    int64     `json:"user_id"`
+	ID        uint64    `json:"id"`
+	UserID    uint64    `json:"user_id"`
 	Name      string    `json:"name"`
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
